Revert "改为float64"

+++ nn/initializer/initializer.go
@@ -1,7 +1,7 @@
 package initializer
 
 type Initializer interface {
-	Rand() float64
-	RandN(n int) []float64
-	RandShape(shapes ...int64) []float64
+	Rand() float32
+	RandN(n int) []float32
+	RandShape(shapes ...int64) []float32
 }
